Some changes related to std::string_view handling

+++ lib/model/log_viewer/LogViewerModel.cpp
@@ -192,7 +192,8 @@ void LogViewerModel::setLogFileName(
     m_isActive = true;
 
     m_currentLogFileStartBytesRead = currentLogFile.read(
-        m_currentLogFileStartBytes.data(), m_currentLogFileStartBytes.size());
+        m_currentLogFileStartBytes.data(),
+        static_cast<qint64>(m_currentLogFileStartBytes.size()));
 
     currentLogFile.close();
